+++ include/xsk/gsc/decompiler.hpp
@@ -15,7 +15,7 @@ class decompiler
     context const* ctx_;
     program::ptr program_;
     decl_function::ptr func_;
-    std::unordered_map<u32, std::string> labels_;
+    std::unordered_map<usize, std::string> labels_;
     std::vector<std::string> expr_labels_;
     std::vector<std::string> tern_labels_;
     std::stack<node::ptr> stack_;
